simple_kv::command_request

Enum RequestData

pub enum RequestData {
    Hget(Hget),
    Hgetall(Hgetall),
    Hmget(Hmget),
    Hset(Hset),
    Hmset(Hmset),
    Hdel(Hdel),
    Hmdel(Hmdel),
    Hexist(Hexist),
    Hmexist(Hmexist),
    Subscribe(Subscribe),
    Unsubscribe(Unsubscribe),
    Publish(Publish),
}

Variants

Hget(Hget)

Hgetall(Hgetall)

Hmget(Hmget)

Hset(Hset)

Hmset(Hmset)

Hdel(Hdel)

Hmdel(Hmdel)

Hexist(Hexist)

Hmexist(Hmexist)

Subscribe(Subscribe)

Unsubscribe(Unsubscribe)

Publish(Publish)

Implementations

impl RequestData

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<RequestData>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.